backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
djoser==2.1.0
setuptools==68.0.0
psycopg2-binary==2.9.10
Pillow
PyYAML
python-dotenv==1.0.1
django-filter==21.1
django-extensions==3.1.3